rust: do not run separate build/install steps
Upstream does not actually use or test it this way; if the goal
is to install items, then install target should be executed directly.

In particular, in latest rust release building stage 2 items has regressed
altogether (incorrect dependencies between rust-analyze tool and the
libs it needs) and no one noticed.

Signed-off-by: Alexander Kanavin <alex@linutronix.de>